requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
pyttsx3
python-dotenv>=1.0.0
google-genai>=1.0.0
